Refs. #47 - Refatoração e ajustes das classes relacionadas a Quesito

Related work items: #47

+++ src/main/java/com/sglp/sglp_api/domain/service/QuesitoService.java
@@ -1,9 +1,10 @@
 package com.sglp.sglp_api.domain.service;
 
 import com.sglp.sglp_api.domain.exception.QuesitoNaoEncontradoException;
+import com.sglp.sglp_api.domain.model.LaudoPericial;
 import com.sglp.sglp_api.domain.model.Quesito;
 import com.sglp.sglp_api.domain.repository.QuesitoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,28 +12,27 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@AllArgsConstructor
 public class QuesitoService {
 
-    @Autowired
-    private QuesitoRepository quesitoRepository;
+    public static final String QUESITO_NAO_ENCONTRADO = "Quesito não encontrado no laudo de ID %s";
 
+    private final QuesitoRepository quesitoRepository;
+    private final LaudoPericialService laudoPericialService;
+    private final QuesitoTransactionalService transactionalService;
 
-    public List<Quesito> listar() {
-        return quesitoRepository.findAll();
+    public List<Quesito> listar(String laudoId) {
+        LaudoPericial laudoPericial = laudoPericialService.buscarOuFalhar(laudoId);
+        return laudoPericial.getQuesitos();
     }
 
-    public Quesito buscar(String quesitoId) {
-        Optional<Quesito> obj = quesitoRepository.findById(quesitoId);
-
-        if(obj.get().getId().equals(quesitoId)){
-            return obj.get();
-        }
-        return null;
+    public Optional<Quesito> buscar(String quesitoId) {
+        return quesitoRepository.findById(quesitoId);
     }
 
     @Transactional
-    public Quesito salvar(Quesito quesito) {
-        return quesitoRepository.save(quesito);
+    public Quesito salvar(String laudoId, Quesito quesito) {
+        return transactionalService.salvar(laudoId, quesito);
     }
 
     public Quesito buscarOuFalhar(String quesitoId) {
@@ -40,7 +40,30 @@ public class QuesitoService {
                 .orElseThrow(() -> new QuesitoNaoEncontradoException(quesitoId));
     }
 
-    public void remover(String quesitoId) {
-        quesitoRepository.deleteById(quesitoId);
+    @Transactional
+    public void remover(String laudoId, String quesitoId) {
+        LaudoPericial laudo = laudoPericialService.buscarOuFalhar(laudoId);
+        List<Quesito> quesitos = laudo.getQuesitos();
+
+        if (quesitos.removeIf(q -> q.getId().equals(quesitoId))) {
+            laudoPericialService.atualizar(laudoId, laudo);
+            quesitoRepository.deleteById(quesitoId);
+        } else {
+            throw new QuesitoNaoEncontradoException(QUESITO_NAO_ENCONTRADO, laudoId);
+        }
+    }
+
+    @Transactional
+    public Quesito atualizar(String laudoId, String quesitoId, Quesito quesito) {
+        Optional<LaudoPericial> laudoOpt = laudoPericialService.buscar(laudoId);
+        Quesito quesitoExistente = buscarOuFalhar(quesitoId);
+
+        if (laudoOpt.isPresent()) {
+            quesitoExistente.setParte(quesito.getParte());
+            quesitoExistente.setPergunta(quesito.getPergunta());
+            quesitoExistente.setResposta(quesito.getResposta());
+            return transactionalService.salvar(laudoOpt.get().getId(), quesitoExistente);
+        }
+        return null;
     }
 }
